+++ crates/analyzer-abstractions/src/futures_extensions/async_extensions.rs
@@ -28,6 +28,9 @@ impl AsyncWork {
 
 impl ArcWake for AsyncWork {
 	fn wake_by_ref(arc_self: &Arc<Self>) {
+		if arc_self.sender.is_closed() {
+			return;
+		}
 		let cloned = arc_self.clone();
 
 		arc_self.sender.send_blocking(cloned).unwrap();
